plataforma especial

+++ arkanoid/arkapoob/src/aplicacion/PlataformaEspecial.java
@@ -2,7 +2,7 @@ package aplicacion;
 
 public class PlataformaEspecial extends Plataforma {
 
-	public PlataformaEspecial(int x, int y,Arkapoob tablero) {
+	public PlataformaEspecial(double x,double y,Arkapoob tablero) {
 		super(x, y,tablero);
 	}
 	public void moveRight() {
